+++ CHANGELOG.md
@@ -4,7 +4,20 @@ All notable changes to this project will be documented in this file.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [1.0.5] - NEXT
+## [1.0.6] - Next
+
+**Milestone**: Mainnet(1.0.0.0)
+
+| Package          | Version | Link                                                               |
+| ---------------- | ------- | ------------------------------------------------------------------ |
+| Symbol Bootstrap | v1.0.6  | [symbol-bootstrap](https://www.npmjs.com/package/symbol-bootstrap) |
+
+-   Reward Program Agent `2.0.0` upgrade.
+-   Added `MonitorOnly` reward program.
+-   The `link` and `enrollRewardProgram` commands allow `--customPreset` to avoid password prompt when main private key is not stored in the target folder. 
+-   Symbol rest `3.5.6` upgrade.
+
+## [1.0.5] - May-3-2021
 
 **Milestone**: Mainnet(1.0.0.0)